taller_mecanico/requirements2.txt
Django==5.2.6
# Django REST Framework para la API
# Verifica compatibilidad con Django 5.x
djangorestframework>=3.15,<4

# Nota:
# Este archivo NO incluye mysqlclient para evitar dependencias nativas.
# Con esto podrás ejecutar el proyecto usando SQLite sin instalar MariaDB/MySQL.
# Si deseas usar MariaDB/MySQL, instala adicionalmente:
#   mysqlclient>=2.2,<3
# y las librerías del sistema correspondientes (p. ej., libmysqlclient-dev).
